+++ packages/react/src/index.ts
@@ -65,7 +65,6 @@ export {
   type SetCenterOptions,
   type FitBoundsOptions,
   type PanelPosition,
-  type ProOptions,
   SelectionMode,
   type SelectionRect,
   type OnError,