+++ packages/system/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './attribution';
 export * from './connections';
 export * from './dom';
 export * from './events';